+++ src/Chrysalis.Test/CborSerializationTests.cs
@@ -58,10 +58,10 @@ public class CborSerializerTests
     // [InlineData("84108200581cbfd1651564e3fe4bae43e1a6ebe090ebdcf69ef3aa684867b49219c41a1dcd6500f6", typeof(RegDrepCert))] //RegDrepCert with Anchor Null
     // [InlineData("83118200581cbd1b2f8d71a5890a38946322b93fc003dc5a61873fdc7e7e5856f3e91a1dcd6500", typeof(UnRegDrepCert))] //UnRegDrepCert
     // [InlineData("83128200581c725cceb7823111ea775edd2cfe577eae89b16520856703095b1b28d782784b68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f43617264616e6f4578706c6f7265722f647265702f6d61696e2f6d657461646174612e6a736f6e6c6458208bff6c4cb86f05458075c787cb843f7dffb52f2b7f438037d291d987d71f7738", typeof(UpdateDrepCert))] //UpdateDrepCert
-    [InlineData("a600818258205c433145addae9555c8948cb6e041e05fe4a3379debd0c948d352aebb9e99b2f00018182583901b3326f3bba239194b2407e6a527948a7e6e3e2d199919fa07f4985f0d07820b822b66d682d973239100f9b614ee18f369c27931ab15d54e61b000000388690171d021a0002b149031a08021a2d048183128200581c725cceb7823111ea775edd2cfe577eae89b16520856703095b1b28d782784b68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f43617264616e6f4578706c6f7265722f647265702f6d61696e2f6d657461646174612e6a736f6e6c6458208bff6c4cb86f05458075c787cb843f7dffb52f2b7f438037d291d987d71f77380800", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Conway)
-    [InlineData("a80082825820083d95d8352c349b453400042cc63ac34a589293dd59ccc8b062f079ea5e05ea0082582066ef6e5199ee3d875516e8ceae113e7f309186c45f1e35ef1c6b35dfb8ca5708020183a20058390195e7d4a51aec52dd528a84ad95e411b3f314a49a97714ebbe735c04768838865663d8666a24e66dd5daedb683eb48d07568b126f485d185a01821a001a3c1ba1581ca6b92cdf28af39782e8baecb712298e0bc91787375569b1f0c8fd703a1462443534276311a0132ba27a300583931905ab869961b094f1b8197278cfe15b45cbe49fa8f32c6b014f85a2db2f6abf60ccde92eae1a2f4fdf65f2eaf6208d872c6f0e597cc10b0701821a2406a090a2581c63f947b8d9535bc4e4ce6919e3dc056547e8d30ada12f29aa5f826b8a158209a9ab7431a3db1542009a50fbe1815054c0b8027164dfc7c1ef10fcdbe002b5501581ca6b92cdf28af39782e8baecb712298e0bc91787375569b1f0c8fd703a1462443534276311a2dd67131028201d81858e4d87989d87982581c63f947b8d9535bc4e4ce6919e3dc056547e8d30ada12f29aa5f826b858209a9ab7431a3db1542009a50fbe1815054c0b8027164dfc7c1ef10fcdbe002b55d879824040d87982581ca6b92cdf28af39782e8baecb712298e0bc91787375569b1f0c8fd703462443534276311b0000001166b125ce1a001373c2581cedbf33f5d6e083970648e39175c49ec1c093df76b6e6a0f1473e47761b000000028f200558581c8807fbe6e36b1c35ad6f36f0993e2fc67ab6f2db06041cfa3a53c04a581c554daf3e27e50c8c779713be6ba70f95c35f23f0552a28e5a7a340148258390122aea2da15e494e01767145d48bda16b6d437f1c449823a044193daf299a82ef56311aa10adf04c0072d4870eb9f4d5ff315132434841b741a001312d0021a0005cf0505a1581df196f5c1bee23481335ff4aece32fe1dfa1aa40a944a66d2d6edc9a9a5000b58203a98f88216e5025db28f06e90c56b3def9f35e3ac2beef37aef7f09ace7a60c00d81825820dd325cdb8c4b2d7fbc66d6e04f2abce41e52fbf55807948aed6b9abb726eb699000e81581cedbf33f5d6e083970648e39175c49ec1c093df76b6e6a0f1473e47761283825820c4a540ac2e06c217dd4fb3f39ca3863da394ba134677dafa9b98830ca71d584d03825820b91eda29d145ab6c0bc0d6b7093cb24b131440b7b015033205476f39c690a51f00825820b91eda29d145ab6c0bc0d6b7093cb24b131440b7b015033205476f39c690a51f01", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Conway)
-    [InlineData("a800818258208d84eef32fcacf5792877809c663ceef9d94a1a83f4eb9979e5d068bb14688b00301838258390179e0c0db2fc4786546f2f7d7914cbe899b61776401d78dd21e44b8626c4734f9fa72a547d83adb86c4431df7c78c8519d8238a09b2a5cc7f821a0011b0dea1581c555a99109b2cf170069e84d98b5a94e9ca15a69d5eb80dcd044598b4a1476e616d65343431018258393195f88cd39531c64a8749a625e5d3f9a79cc85d2001db23d851ad65bd19971dfb46a7fa611ddf545ee84dabd741454c4837669de8322d71811a000f42408258390179e0c0db2fc4786546f2f7d7914cbe899b61776401d78dd21e44b8626c4734f9fa72a547d83adb86c4431df7c78c8519d8238a09b2a5cc7f1a048e908c021a0003388d031a0806dede07582052966b4accdf10a1d562d671ffebe13583892a46dcc63fc1be5038702689bd1a081a0806d0ce09a1581c555a99109b2cf170069e84d98b5a94e9ca15a69d5eb80dcd044598b4a1476e616d65343431010e82581c79e0c0db2fc4786546f2f7d7914cbe899b61776401d78dd21e44b862581c6c4734f9fa72a547d83adb86c4431df7c78c8519d8238a09b2a5cc7f", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Conway)
-    [InlineData("a800838258201804b13f34a8452066ff1cad68bd09383f65d407f2f84b31fce0ef7906afe28e00825820fb62b5141e0e6595e4e7382d7b16747785b3333e89d65079c2d39f3d65b1e73100825820fb62b5141e0e6595e4e7382d7b16747785b3333e89d65079c2d39f3d65b1e73101018282583901c63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d22968046e259d10fbd46bf394d0327a32ea71afe3f8910c30beee404821a00165c38a2581c2e8e5e8e1d7837ccc509e30045f656f9e9ebed7ef5621085172d0e9aa446636c617261320146636c617261330146636c617261340147636c617261313101581cd436d9f6b754582f798fe33f4bed12133d47493f78b944b9cc55fd18a15053756d6d69744c6f64676531363336330182583901c63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d22968046e259d10fbd46bf394d0327a32ea71afe3f8910c30beee4041a010465e9021a0007cfe70b5820126750c84767cace394a3bf78350a69d2fbc9e470975a47e73672fb513431dc00d81825820fb62b5141e0e6595e4e7382d7b16747785b3333e89d65079c2d39f3d65b1e731010e81581cc63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d1082583901c63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d22968046e259d10fbd46bf394d0327a32ea71afe3f8910c30beee4041a00e55388111a000bb7db", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Babbage)
+    // [InlineData("a600818258205c433145addae9555c8948cb6e041e05fe4a3379debd0c948d352aebb9e99b2f00018182583901b3326f3bba239194b2407e6a527948a7e6e3e2d199919fa07f4985f0d07820b822b66d682d973239100f9b614ee18f369c27931ab15d54e61b000000388690171d021a0002b149031a08021a2d048183128200581c725cceb7823111ea775edd2cfe577eae89b16520856703095b1b28d782784b68747470733a2f2f7261772e67697468756275736572636f6e74656e742e636f6d2f43617264616e6f4578706c6f7265722f647265702f6d61696e2f6d657461646174612e6a736f6e6c6458208bff6c4cb86f05458075c787cb843f7dffb52f2b7f438037d291d987d71f77380800", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Conway)
+    // [InlineData("a80082825820083d95d8352c349b453400042cc63ac34a589293dd59ccc8b062f079ea5e05ea0082582066ef6e5199ee3d875516e8ceae113e7f309186c45f1e35ef1c6b35dfb8ca5708020183a20058390195e7d4a51aec52dd528a84ad95e411b3f314a49a97714ebbe735c04768838865663d8666a24e66dd5daedb683eb48d07568b126f485d185a01821a001a3c1ba1581ca6b92cdf28af39782e8baecb712298e0bc91787375569b1f0c8fd703a1462443534276311a0132ba27a300583931905ab869961b094f1b8197278cfe15b45cbe49fa8f32c6b014f85a2db2f6abf60ccde92eae1a2f4fdf65f2eaf6208d872c6f0e597cc10b0701821a2406a090a2581c63f947b8d9535bc4e4ce6919e3dc056547e8d30ada12f29aa5f826b8a158209a9ab7431a3db1542009a50fbe1815054c0b8027164dfc7c1ef10fcdbe002b5501581ca6b92cdf28af39782e8baecb712298e0bc91787375569b1f0c8fd703a1462443534276311a2dd67131028201d81858e4d87989d87982581c63f947b8d9535bc4e4ce6919e3dc056547e8d30ada12f29aa5f826b858209a9ab7431a3db1542009a50fbe1815054c0b8027164dfc7c1ef10fcdbe002b55d879824040d87982581ca6b92cdf28af39782e8baecb712298e0bc91787375569b1f0c8fd703462443534276311b0000001166b125ce1a001373c2581cedbf33f5d6e083970648e39175c49ec1c093df76b6e6a0f1473e47761b000000028f200558581c8807fbe6e36b1c35ad6f36f0993e2fc67ab6f2db06041cfa3a53c04a581c554daf3e27e50c8c779713be6ba70f95c35f23f0552a28e5a7a340148258390122aea2da15e494e01767145d48bda16b6d437f1c449823a044193daf299a82ef56311aa10adf04c0072d4870eb9f4d5ff315132434841b741a001312d0021a0005cf0505a1581df196f5c1bee23481335ff4aece32fe1dfa1aa40a944a66d2d6edc9a9a5000b58203a98f88216e5025db28f06e90c56b3def9f35e3ac2beef37aef7f09ace7a60c00d81825820dd325cdb8c4b2d7fbc66d6e04f2abce41e52fbf55807948aed6b9abb726eb699000e81581cedbf33f5d6e083970648e39175c49ec1c093df76b6e6a0f1473e47761283825820c4a540ac2e06c217dd4fb3f39ca3863da394ba134677dafa9b98830ca71d584d03825820b91eda29d145ab6c0bc0d6b7093cb24b131440b7b015033205476f39c690a51f00825820b91eda29d145ab6c0bc0d6b7093cb24b131440b7b015033205476f39c690a51f01", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Conway)
+    // [InlineData("a800818258208d84eef32fcacf5792877809c663ceef9d94a1a83f4eb9979e5d068bb14688b00301838258390179e0c0db2fc4786546f2f7d7914cbe899b61776401d78dd21e44b8626c4734f9fa72a547d83adb86c4431df7c78c8519d8238a09b2a5cc7f821a0011b0dea1581c555a99109b2cf170069e84d98b5a94e9ca15a69d5eb80dcd044598b4a1476e616d65343431018258393195f88cd39531c64a8749a625e5d3f9a79cc85d2001db23d851ad65bd19971dfb46a7fa611ddf545ee84dabd741454c4837669de8322d71811a000f42408258390179e0c0db2fc4786546f2f7d7914cbe899b61776401d78dd21e44b8626c4734f9fa72a547d83adb86c4431df7c78c8519d8238a09b2a5cc7f1a048e908c021a0003388d031a0806dede07582052966b4accdf10a1d562d671ffebe13583892a46dcc63fc1be5038702689bd1a081a0806d0ce09a1581c555a99109b2cf170069e84d98b5a94e9ca15a69d5eb80dcd044598b4a1476e616d65343431010e82581c79e0c0db2fc4786546f2f7d7914cbe899b61776401d78dd21e44b862581c6c4734f9fa72a547d83adb86c4431df7c78c8519d8238a09b2a5cc7f", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Conway)
+    // [InlineData("a800838258201804b13f34a8452066ff1cad68bd09383f65d407f2f84b31fce0ef7906afe28e00825820fb62b5141e0e6595e4e7382d7b16747785b3333e89d65079c2d39f3d65b1e73100825820fb62b5141e0e6595e4e7382d7b16747785b3333e89d65079c2d39f3d65b1e73101018282583901c63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d22968046e259d10fbd46bf394d0327a32ea71afe3f8910c30beee404821a00165c38a2581c2e8e5e8e1d7837ccc509e30045f656f9e9ebed7ef5621085172d0e9aa446636c617261320146636c617261330146636c617261340147636c617261313101581cd436d9f6b754582f798fe33f4bed12133d47493f78b944b9cc55fd18a15053756d6d69744c6f64676531363336330182583901c63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d22968046e259d10fbd46bf394d0327a32ea71afe3f8910c30beee4041a010465e9021a0007cfe70b5820126750c84767cace394a3bf78350a69d2fbc9e470975a47e73672fb513431dc00d81825820fb62b5141e0e6595e4e7382d7b16747785b3333e89d65079c2d39f3d65b1e731010e81581cc63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d1082583901c63fe326cf6ca78124768d68c501008c881b132ef83999c99bb8a23d22968046e259d10fbd46bf394d0327a32ea71afe3f8910c30beee4041a00e55388111a000bb7db", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Babbage)
     [InlineData("a500828258209a8fbe63b8d95c30746d3337fafd8fbe11dfa72bf05c3e7876fc28665a745bf900825820548650236f3bae15b8ddb7d65dde5ba6fb1785606880f44eff04b2316f6affde0101828258390109c230576e5d6c062a0fa95edafb4b7760bf48c9569c949294993befdcffc654e5f87793e8875d6cb40cfc8da22aeeecd7b62dc26e037df5821a00169b08a1581c49450fe5ce314aa6ef5d4c947d69d4af8e298a005114718a7cc1f784a14d41444150756e6b334430333530018258390109c230576e5d6c062a0fa95edafb4b7760bf48c9569c949294993befdcffc654e5f87793e8875d6cb40cfc8da22aeeecd7b62dc26e037df51a04144c52021a000fa6b90b5820318f56b89c98d3785d39da7f7b5e519843101a7e6256ade04d10ec8ee695153f0d81825820b7aeaf10a7addfcf7a0011c5520b393d3f7450fa9dcb8d7528c565737499cb0c00", typeof(TransactionBody))] // Serialized CBOR for TransactionBody(Alonzo)
     public void SerializeAndDeserializePrimitives(string cborHex, Type type)
     {
