+++ src/Chrysalis/Cardano/Models/Core/Script/PlutusScript.cs
@@ -1,7 +1,7 @@
 using Chrysalis.Cardano.Models.Cbor;
 using Chrysalis.Cbor;
 
-namespace Chrysalis.Cardano.Models.Core;
+namespace Chrysalis.Cardano.Models.Core.Script;
 
 [CborSerializable(CborType.Bytes)]
 public record PlutusV1Script(CborBytes Value): ICbor;
