+++ src/Chrysalis/Cardano/Models/Core/Transaction/TransactionWitnessSet.cs
@@ -6,12 +6,13 @@ namespace Chrysalis.Cardano.Models.Core.Transaction;
 
 [CborSerializable(CborType.Map)]
 public record TransactionWitnessSet(
-    [CborProperty(0)] CborDefiniteList<VKeyWitness>? VKeyWitnessSet,
-    [CborProperty(1)] CborDefiniteList<NativeScript>? NativeScriptSet,
-    [CborProperty(2)] CborDefiniteList<BootstrapWitness>? BootstrapWitnessSet,
-    [CborProperty(3)] CborDefiniteList<BootstrapWitness>? PlutusV1ScriptSet, // @TODO: Modify T Parameter    
-    [CborProperty(4)] CborDefiniteList<BootstrapWitness>? PlutusDataSet, // @TODO: Modify T Parameter   
+    [CborProperty(0)] CborIndefiniteList<VKeyWitness>? VKeyWitnessSet,
+    [CborProperty(1)] CborIndefiniteList<NativeScript>? NativeScriptSet, // @TODO: Modify T Parameter
+    [CborProperty(2)] CborIndefiniteList<BootstrapWitness>? BootstrapWitnessSet,
+    [CborProperty(3)] CborIndefiniteList<PlutusV1Script>? PlutusV1ScriptSet,   
+    [CborProperty(4)] CborIndefiniteList<BootstrapWitness>? PlutusDataSet, // @TODO: Modify T Parameter   
     [CborProperty(5)] Redeemers? Redeemers,
-    [CborProperty(6)] CborDefiniteList<BootstrapWitness>? PlutusV2ScriptSet, // @TODO: Modify T Parameter
-    [CborProperty(7)] CborDefiniteList<BootstrapWitness>? PlutusV3ScriptSet // @TODO: Modify T Parameter
+    [CborProperty(6)] CborIndefiniteList<PlutusV2Script>? PlutusV2ScriptSet,
+    [CborProperty(7)] CborIndefiniteList<PlutusV3Script>? PlutusV3ScriptSet
+    
 ) : ICbor;
